worker/requirements.txt
celery==5.3.4
python-dotenv==1.0.0
pymongo==4.5.0
ftfy
regex
tqdm
torch
torchvision